+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - replaced Spectron with playwright
 - Changed release build on MacOS 10.15
+- Fixed Colour Palette to close modal
+- Colour Palette clear button in danger zone.
+- Colour Palette buttons relabeled
+- Advert no longer an iframe
 
 ## 0.1.24 (2022-12-19)